config auth jwt oauth github

+++ src/routes/memories.ts
@@ -3,8 +3,17 @@ import { z } from 'zod'
 import { FastifyInstance } from 'fastify'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
+    const { sub } = request.user
+
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -17,7 +26,9 @@ export async function memoriesRoutes(app: FastifyInstance) {
     }))
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
+    const { sub } = request.user
+
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -28,14 +39,21 @@ export async function memoriesRoutes(app: FastifyInstance) {
       where: { id },
     })
 
+    if (!memory.isPublic && memory.userId !== sub) {
+      return reply.status(401).send()
+    }
+
     return memory
   })
 
   app.post('/memories', async (request) => {
+    const { sub } = request.user
+
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
       isPublic: z.boolean().default(false),
+      userId: z.string(),
     })
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
@@ -45,14 +63,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '',
+        userId: sub,
       },
     })
 
     return memory
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
+    const { sub } = request.user
+
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -67,7 +87,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: { id },
       data: {
         content,
@@ -79,14 +109,26 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
+    const { sub } = request.user
+
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
 
-    const memory = await prisma.memory.delete({
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== sub) {
+      return reply.status(401).send()
+    }
+
+    await prisma.memory.delete({
       where: { id },
     })
